[#1107] Feature/ Add CLA Manager Request
- Added swagger spec for CLA Manager request endpoint
