+++ cla-backend-go/swagger/cla.v2.yaml
@@ -1123,6 +1123,46 @@ paths:
       tags:
         - signatures
 
+  /company/{companySFID}/project/{projectSFID}/claGroup/{projectID}/cla-manager/requests:
+    post:
+      summary: Adds a CLA Manager Designee to the specified Company and Project
+      description: User proposes a CLA Manager making the proposed user CLA Manager Designee
+      operationId: createCLAManagerRequest
+      parameters:
+        - $ref: "#/parameters/x-request-id"
+        - $ref: "#/parameters/x-acl"
+        - $ref: "#/parameters/x-username"
+        - $ref: "#/parameters/x-email"
+        - $ref: "#/parameters/path-companySFID"
+        - $ref: "#/parameters/path-projectSFID"
+        - $ref: "#/parameters/path-projectID"
+        - name: body
+          in: body
+          schema:
+            type: object
+            x-nullable: false
+            title: CLA Manager Designee User
+            description: User object used for CLA Manager requests
+            properties:
+              userEmail:
+                type: string
+              contactAdmin:
+                type: boolean
+      responses:
+        '200':
+          description: 'Success'
+          schema:
+            $ref: "#/definitions/cla-manager-designee"
+        '400':
+          $ref: '#/responses/invalid-request'
+        '403':
+          $ref: '#/responses/forbidden'
+        '409':
+          $ref: '#/responses/conflict'
+      tags:
+        - cla-manager
+            
+
   /company/{companySFID}/project/{projectSFID}/claGroup/{projectID}/cla-manager:
     post:
       summary: Adds a new CLA Manager to the specified Company and Project
